requirements.txt
aiocoap==0.4.3
paho-mqtt==1.6.1
requests==2.25.1
cachetools==5.0.0
Flask==2.0.2
pymongo==4.0.1